Refactor CashierSessionController and add ProductController
- Updated CashierSessionController to enhance session handling, including improved cart data retrieval and total calculations.
- Introduced ProductController to serve product details from a JSON file.
- Added a new JSON data file for product information.
- Enhanced OrderSummary and related components to support new session summary and close session modals.
- Updated tax rate sharing in HandleInertiaRequests middleware.
