+++ resources/data/products.json
@@ -1,0 +1,110 @@
+{
+    "pos-terminal": {
+        "name": "QuickJuan POS Terminal",
+        "image": "/images/resto.png",
+        "slug": "pos-terminal",
+        "description": "Comprehensive point of sale solution for restaurants and retail businesses.",
+        "capabilities": [
+            {
+                "title": "Flexible Payments",
+                "desc": "Accept cash, card, QR, and e-wallets with ease."
+            },
+            {
+                "title": "Inventory Management",
+                "desc": "Track stock in real time and get low-inventory alerts."
+            },
+            {
+                "title": "Sales Analytics",
+                "desc": "Visualize sales trends and make data-driven decisions."
+            },
+            {
+                "title": "Customer Database",
+                "desc": "Build loyalty with customer profiles and purchase history."
+            },
+            {
+                "title": "Fast Checkout",
+                "desc": "Speedy transactions keep lines moving and customers happy."
+            },
+            {
+                "title": "Smart Notifications",
+                "desc": "Get notified for important events and low stock."
+            },
+            {
+                "title": "Digital Receipts",
+                "desc": "Send receipts via SMS or email instantly."
+            },
+            {
+                "title": "Cloud Access",
+                "desc": "Manage your business from anywhere, anytime."
+            },
+            {
+                "title": "Integrations",
+                "desc": "Connect with accounting, delivery, and more."
+            }
+        ]
+    },
+    "mobile-pos": {
+        "name": "QuickJuan Mobile POS",
+        "image": "/images/resto2.png",
+        "slug": "mobile-pos",
+        "description": "Mobile point of sale system for on-the-go transactions and flexibility.",
+        "capabilities": [
+            {
+                "title": "Mobile Checkout",
+                "desc": "Sell anywhere with your phone or tablet."
+            },
+            {
+                "title": "All-day Battery",
+                "desc": "Optimized for long shifts and mobility."
+            },
+            {
+                "title": "Inventory Sync",
+                "desc": "Real-time inventory updates across devices."
+            },
+            {
+                "title": "Digital Receipts",
+                "desc": "Send receipts via SMS or email instantly."
+            },
+            {
+                "title": "Instant Alerts",
+                "desc": "Get notified for important events and low stock."
+            },
+            {
+                "title": "Cloud Access",
+                "desc": "Manage your business from anywhere, anytime."
+            }
+        ]
+    },
+    "kiosk": {
+        "name": "QuickJuan Kiosk",
+        "image": "/images/resto3.png",
+        "slug": "kiosk",
+        "description": "Self-service kiosk for efficient customer ordering and payment processing.",
+        "capabilities": [
+            {
+                "title": "Self-Service",
+                "desc": "Let customers order and pay on their own."
+            },
+            {
+                "title": "Digital Receipts",
+                "desc": "Send receipts via SMS or email instantly."
+            },
+            {
+                "title": "Secure Payments",
+                "desc": "PCI-compliant and safe for all transactions."
+            },
+            {
+                "title": "Inventory Sync",
+                "desc": "Real-time inventory updates across devices."
+            },
+            {
+                "title": "Sales Analytics",
+                "desc": "Visualize sales trends and make data-driven decisions."
+            },
+            {
+                "title": "Cloud Access",
+                "desc": "Manage your business from anywhere, anytime."
+            }
+        ]
+    }
+}
